+++ docker/Dockerfile
@@ -29,8 +29,6 @@ USER root
 RUN apt-get update && apt-get install --no-install-recommends --yes \
         gcc \
         gphoto2 \
-        ncurses-dev \
-        readline-common \
         udev
 
 # Install program dependencies.
@@ -39,13 +37,15 @@ WORKDIR ${POCS}
 RUN mkdir -p "${PANDIR}/scripts" && \
     cd "${PANDIR}/scripts" && \
     # Install gphoto2 auto-updater
-    wget $gphoto2_url -O gphoto2-updater.sh && \
-    chmod +x gphoto2-updater.sh && \
+#    wget $gphoto2_url -O gphoto2-updater.sh && \
+#    chmod +x gphoto2-updater.sh && \
     # Don't actually update gphoto2 as of right now.
-    # sudo /bin/bash gphoto2-updater.sh --stable && \
+#    sudo /bin/bash gphoto2-updater.sh --stable && \
+#    rm gphoto2-updater.sh && \
     # Install arduino-cli.
     wget -q "${arduino_url}" -O install-arduino-cli.sh && \
     sudo BINDIR="/usr/local/bin" /bin/sh install-arduino-cli.sh && \
+    rm install-arduino-cli.sh && \
     sudo chown -R "${PANUSER}":"${PANUSER}" "${PANDIR}"
 
 # Install the module.
