+++ src/panoptes/pocs/base.py
@@ -29,13 +29,13 @@ class PanBase(object):
 
         self.logger = get_logger()
 
-        # If the user requests a db_type then update runtime config
-        db_type = kwargs.get('db_type', self.get_config('db.type', default='file'))
-        db_name = kwargs.get('db_name', self.get_config('db.name', default='panoptes'))
-        db_folder = kwargs.get('db_folder', self.get_config('db.folder', default='json_store'))
-
         global PAN_DB_OBJ
         if PAN_DB_OBJ is None:
+            # If the user requests a db_type then update runtime config
+            db_type = kwargs.get('db_type', self.get_config('db.type', default='file'))
+            db_name = kwargs.get('db_name', self.get_config('db.name', default='panoptes'))
+            db_folder = kwargs.get('db_folder', self.get_config('db.folder', default='json_store'))
+
             PAN_DB_OBJ = PanDB(db_type=db_type, db_name=db_name, storage_dir=db_folder)
 
         self.db = PAN_DB_OBJ
