+++ src/panoptes/pocs/camera/gphoto/__init__.py
@@ -1,0 +1,1 @@
+from panoptes.pocs.camera.gphoto.base import AbstractGPhotoCamera
