--- src/panoptes/pocs/camera/simulator/__init__.py
@@ -1,1 +1,0 @@
-from panoptes.pocs.camera.simulator.dslr import Camera
