--- src/panoptes/pocs/camera/simulator_sdk/__init__.py
@@ -1,1 +1,0 @@
-from panoptes.pocs.camera.simulator_sdk.ccd import Camera
